+++ RELEASE_NOTES.md
@@ -4,11 +4,12 @@ This file contains one line summaries of commits that are worthy of mentioning i
 A new header is inserted each time a *tag* is created.
 
 ## main branch
+ 
 - backend: add support for GGP platform
-
 - gltfio: calculate primitive's AABB correctly. 
 - gltfio: fix glTF breaking issue
 - engine: primitives with `CullingMode::FRONT_AND_BACK` are now skipped.
+- engine: add missing getters on `MaterialInstance`
 
 ## v1.28.2